update: eloRating to BigDecimal for precision in calculations

+++ src/main/java/com/emt/model/response/PlayerResponse.java
@@ -1,8 +1,9 @@
 package com.emt.model.response;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import lombok.Builder;
 
 @Builder
 public record PlayerResponse(
-    Long playerId, String nickname, Integer eloRating, Instant registeredAt) {}
+    Long playerId, String nickname, BigDecimal eloRating, Instant registeredAt) {}
